+++ tests/runTests.ts
@@ -3,6 +3,7 @@ import "./renderVisibility.test";
 import "./gridUtils.test";
 import "./fontData.test";
 import "./fontPersistence.test";
+import "./migrationScripts.test";
 import "./localStorageUtils.test";
 import "./exportUtils.test";